+++ ChangeLog.md
@@ -4,6 +4,7 @@
 - BugillaServer can now be fully qualified
 - Add apikeySession to support api_key auth
 - Add isNotEmpty search expression operator
+- Add changedSince and changedUntil search expression
 - Add searchBugsAllWithLimit, searchBugsAll and getBugAll to get all the bug fields
 - Change Bug to include ExternalBugs information
 - Export Request